+++ admin/partials/decalog-admin-view-traces.php
@@ -22,7 +22,7 @@ $traceListTable->prepare_items();
 <div class="wrap">
 	<h2><?php echo sprintf( esc_html__( '%s Traces Viewer', 'decalog' ), DECALOG_PRODUCT_NAME );?></h2>
 	<?php $traceListTable->views(); ?>
-	<form id="events-filter" method="get" action="<?php echo admin_url('admin.php'); ?>">
+	<form id="traces-filter" method="get" action="<?php echo admin_url('admin.php'); ?>">
 		<input type="hidden" name="page" value="decalog-tviewer" />
 		<?php $traceListTable->display(); ?>
 	</form>