+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - DecaLog now supports metrics collecting, forwarding and publishing.
 - DecaLog now supports traces collecting and publishing.
+- New traces viewer allowing to view traces right in the admin dashboard. 
 - New "metrics" tab in settings to list all currently collected metrics.
 - DecaLog now collects extended metrics for: PHP, WordPress core, database, WordPress multisite and plugins / themes using [WordPress DecaLog SDK](https://github.com/Pierre-Lannoy/wp-decalog-sdk).
 - DecaLog now collects minimalist metrics for Wordfence.
@@ -21,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - New logger to send traces to a Zipkin instance.
 - New logger to send traces to Grafana Cloud via Grafana agent.
 - New logger to send traces to Datadog via Datadog agent.
+- New logger to store traces in WordPress database or with APCu.
 - [WP-CLI] New command to list, dump or get metrics: see `wp help log metrics` for details.
 - [WP-CLI] New command to activate/deactivate endpoint authentication: see `wp help settings` for details.
 - [BC] The version of DecaLog API is now `v3`.